Create DrinkMachineRev.01.ino
This can be loaded to an Arduino to take a serial input (ESP8266 or USB) and mix drinks up to 5 in parallel (in its current form) via peristaltic pumps into a glass. This includes glass detection, glass full and glass removed during pour capability so as not to keep pouring if the glass is removed or to start pouring if there is no glass to begin with. If the glass is full/not removed since last pour, new drinks will not pour, preventing over fill.